Separate Notion client wrapper w/naive conversion of page titles to <h1>
Note: although not displayed by Notion, italics, bold, etc. are still used behind the scenes…presumably in caes one moves some formatted text to a page title and back again? Anyway, here we use a naïve bit of decoding to extract and recombine page titles.

Also tweaked the interface to the `log` function so it's harder to misuse