Changes to npm scripts: `npm start` + `npm run run-once`
[Irrelevant sidenote: I wish that npm had a default behavior of assuming you meant `run x` if you feed it a command that isn't part of its built-in library. In other words, I should be able to enter `npm run-once` and have it, upon seeing that such a script exists, trigger that script. 😤]

# Highly Relevant Notes

## Requirements

As I'll elaborate on in a moment, this is a highly idiosyncratic addition, largely intended to increase my current productivity with some thought — but not much! — put into portability. I _think_ this setup should work

### Fish Shell

You don't have to switch to `fish` as your primary shell, but it's harmless to install, it's really just another incredible free tool you can shove into the Mary Poppins Bag that is your dev system, just in case. Who knows, maybe you'll just give it a go some time…having installed it for `notion2svelte`, you'll type _fish_ into your terminal, find that it's got some lovely defaults, and end up marrying it in the Church of Default Shell. But for now, you're probably just wondering _why_ I chose to use `fish` in this `watch` script, right?

Why, indeed. Honestly, I'm just lazy, so I'm using my default shell. If you're reading this* and thing it should work differently, let me know and we'll talk.

### `entr`

This is the simplest file-watching tool I could find and it worked more or less on the first try. ¯\_(ツ)_/¯

_In short_…once you [install entr](https://github.com/eradman/entr) and [install fish], `npm run watch` should allow you to make changes to the pipeline and see them instantly reflected in your running svelte app. As a bonus, `entr` lets you hit <spacebar> to manually trigger the script if, for instance, you've made changes to a Notion page and so you want to re-pull ready-to-publish pages. It doesn't sound cool when I describe it, but it's crazy magic yumminess in practice!🍦

---

*HOW!? I haven't made this repo public yet, so either you've hacked my shit (curse you!!!) or…you're from (hushed tones) The Future! How is it? Still no flying cars, I'm guessing? Ah, well. You can't win 'em all.