Images now stored locally
# Discussion

Notion hosts images on AWS with a very short expiry on each image URL.

If I'm reading things rigthly, Notion's servers have a private token that they can use to refetch an expired image, but, since we don't have access to that token, it's best that we get it out of that locked ecosystem.

This is a pretty naive way to do that, but…baby steps, neh?

# Details

- Add a `OUTPUT_STATIC_FOLDER=` pointing to Svelte Kit's `<svelte-root>/static/` folder (unless you're using `<svelte-root>/public/`, or something even more customized)
- That should be all that's necessary. These updates use `curl` to download each image (in sequence…so…plenty of room for optimization later!). Here's a quick example:

Say I have this directory structure in my target app:

```
<svelte-root>/
  - static/
  - src/
  …
```

Now I process two pages, whose respective slugs are, creatively, "a-page" and "another-page", each of which has two image blocks. The output will should look like this, where each opaque ID is unique:

```
<svelte-root>/
  - static/
    - images/
      - a-page/
        <opaque-block-id>.jpg
        <opaque-block-id>.jpg
      - another-page/
        <opaque-block-id>.jpg
        <opaque-block-id>.jpg
  - src/
  …
```

The URL for each image will be automatically generated for the Image component, of course, so that it shows up in your local page. URLs will come out like:

```
http://<domain>/images/a-page/<opaque-block-id>.jpg
http://<domain>/images/another-page/<opaque-block-id>.jpg
```

Images still don't support any sort of custom sizing or placement, but that may come later in the shape of a caption-embedded `Ro††en` hint. TBD