✅ Divider components
Note that the simplest option here is just `<hr />` from sveltifier plus CSS in the target project. This demonstrates how easily we can swap in…well, wtf we want. After all, the Divider component could take further input from the app in order to change dynamically. Not sure why you'd want to, but…sky's the limit! 😬