Fixed: callstack exceeded error
The problem: I had <Magic> taking in the page object -- for every block, mind you! -- in its entirety, _including_ all the sub-blocks. This meant that, for everby individual block, I was adding every block AGAIN. This doesn't show up in the output because I happened to chose an implementation of Magic that doesn't use the page JSON, so the excess compiles away. On a page with only, say, 3 blocks, this meant making the compiler handle 9 chunks of redundant block JSON. With the Bells & Whistles test page, which has lots of blocks and some sub-pages to-boot, the result was a test.svelte file just short of 140,000 lines long. 😳

(As you can see, it's still over 7500 lines long, so this strategy is certainly not one that everyone's going to want to use. I myself will be using it to add comment/feedback widgets to everything I publish. At least, that's the original motivation for doing this less complex Magic demo.)