Tweak to InlineCode styling
Motivation: I'd accidentally rendered an SVG in the middle of some text and came to the conclusion that the system's fine as-is for now as I can just use `&lt;Divider />` insetad of `<Divider />` to get things rendering the way I want. (This is demonstrated with the new content in test.svelte (as of this commmit, obvs). During the investigation, I tried wrapping the slot content in <pre> and found that this results in an interesting crossing-boxes effect, which I decided to keep, albeit with a bit of aesthetic tweaking and background gradient darkening