Route change
This is an experiment to see what happens when a page's slug changes in Notion. In  order to get git to see the resulting output as a rename, I had to manually `rm bells-and-whistles.*`, as well as `assets/bells-and-whistles/`. This might seem like something worth automating away, but I actually like it. Route changes should always be carefully considered since SEO and user bookmarks can be affected, so a certain amount of built-in friction can't hurt!