Removed hard-coded `alt` tag
Though it annoys me to have all notion2svelte images trigger a Svelte warning by default, I can't think of a default that's better than the warning, so we'll follow the Svetle team's lead and just assume this is the reader's job to resolve. (I've ideas for how notion2svelte could use a proprietary 'tag" in the image caption as a way to overcome Notion's limitations here, but they're not likely to hit the top of the Product Backlog (aka DonenessTracker) any time soon.