Added <Magic> wrapper to title, as well
(And, I think, fixed a bug caused by sveltifier.js:168…? `notionContext` seemed to be the wrong value to pass in. I don't know how I could have let such a boneheaded error go for so long, so it's possible that I just broke something that was woking. sigh)

Note: As always, changes in demo-app are technically not part of _notion2svelte_. That said, I like the changes I made to Embed.svelte & Magic.svelte's "inspection" view. Specifically, the JSON starts out hidden behind a summary/details toggle, and a new, additional toggle contains a "pathified" view of the JSON. I find this makes it easier to know which exact path one is looking for when tanslating the JSON to code, so hopefully this gives a nice speed bump to my future work with the Notion API.