Added multi-regional timestamp to make it easier to subliminally track notion2svelte runs in Terminal
This is especially true when using `npm start`, since that command responds to file changes, and I've got VS Code set to Auto Save: onWindowChange, which means it's not hard to accidentally trigger a rerun of notion2svelte. The resultant output can come out so fast that one doesn't even see it happen unless Terminal is always unhidden by other windows. This change makes it more obvious when that's happening.