Tighter code styles
Problem: Code in a two-column layout looked shite because the "main width" used here (that calc thing with all the notes about duplication) carries fewer characters per line than Notion's design. Shrinking the text and reducing whitespace helps in this case.

Of course, so would incorporating Highlight.js or the like, but we're keeping it simple for now!