🎪 Big ol' change <Magic> → blockProps
- Removed all traces of Magic component, both from demo-app and _notion2svelte_
- Implemented blockProps on Paragraphs (only)
  - In this implementation, a Notion-ish now control appears when hovering paragraphs
  - When clicked, the control displays the raw blockProps, along with an "Open block in Notion" button, whose funtionality in turn depends on blockProps

Note the new "m.svelte" page. I think I'll use this in general for testing new things and/or doing the initial replication on bugs. The crazy-shorts slug ("m") is a good reminder to KEEP IT TO A MINIMUM! By which I mean, _notion2svelte_ can get confusing to troubleshoot when more than one thing is broken, so, when investigating the unknown, it helps to cut out as much noise as possible!