Clean slate for the docs as I near completion
This is probably a horrible idea that will bloat the repo with a bunch more binaries from the various screenshots and demo images in `assets`. If so, som advanced git fu may be in order to clear the history…but first, some better solution has to be found for handling Notion's ephemeral image URLs!