Round 1 of final documentation
Just for kicks, I decided to use Synced Blocks for the list of docs at the bottom (which doubles as the navigation). It's a probably-horrible idea, but it's enough to get me out the door without having to invent something new, so…¯\_(ツ)_/¯

One obvious downside: every page in the navigation already needs to exist in the _notion2svelte_-rendered local files, else the navigation links will render as errors. (This is a side effect of decoupling _notion2svelte_'s output from Notion: we generate a local mapping of page ids↔slugs so that we can generate internal links without a return trip to Notion.)

My solution, for now, is creating skeleton pages for every doc I plan to write so that the nav can be rendered correctly as I work my way through writing and notion2svelte-ing the remaining docs, one by one.

Anyway, in this commit, these pages can be considered final*:

- the Untitled home page (about-notion2svelte.svelte)
- Components
- Writing Good Slugs
- blockProps
- Nested Blocks
- Synced Blocks

*Pending discovery of any errors