pure refactor - quick re-save to ensure auto-format's universally applied
As much as I'm grateful for Prettier, inasmuch as it makes it easier to _approach_ the joy of elm-format, it still falls short inasmuch as it has configuration options. I'd much rather have sensible, unalterable defaults as it avoids this sort of nonsense commit.

In case you've somehow stumbled upon this text (which I fully expect no one to ever read again), know this: it's a pain in the ass to start working in a file, and only realize after making a handful of changes that the changes themselves are lost to git in a sea of autoformat updates! 😤

🤓