Simplified InlineCode w/minor style tweaks
The original goal was to just remove the padding as I decided it was obnoxious having a gap between the visible boundary of inline code spans and surrounding punctuation—`when rendering something like this`, for example. I then realized that the primary reason to even have the outer `<span>` was the now-removed padding, so I removed that span as well. As a result:

- `font-family` now has an actual effect. Turns out the span's font-family was getting ignored.
- `font-size` is no longer 85% of 85% (applied at two levels), but a single 72% equivalent